Started using C-blocks (note that they should not be used in ObjC or sensitive places like dispatcher.cpp)
All this was to start cleaning-up string code.
I removed the function pointers for each encoding as I'm pretty sure no
one will use them to extend the encoding-handling and they make the code
harder to maintain.

Feature-wise currently the only change is that String#inspect is much
better when part of a string is invalid:
% ./miniruby -e 'p "あ\xFF"'
"あ\xFF"

git-svn-id: http://svn.macosforge.org/repository/ruby/MacRuby/trunk@5049 23306eb0-4c56-4727-a40e-e92c0eb68959