changed the internal representation of strings
The strings could have 2 internal representations (UTF-16 or binary),
there is now only the binary one. It makes a few things harder, but
other things much simpler.

The main reason for doing it is that we could have problems in
multi-threaded applications, when multiple threads are using the same
string at the same time, even without using any operation modifiying the
string (as some operations were prefering using the string in UTF-16 and
others in binary mode).

git-svn-id: http://svn.macosforge.org/repository/ruby/MacRuby/trunk@5054 23306eb0-4c56-4727-a40e-e92c0eb68959