Make Dispatch::Queue#apply documentation less misleading
Claiming that the blocks are executed in parallel is not correct
since it depends on the type of queue being used, and even in that
case concurrency is not the same things as parallelism.

In the example, the array grows dynamically and the assignment
being performed in the block would not be thread-safe
(concurrency-safe?).