Address GH-90 by breaking out Xcode integration installation
Xcode integration is now installed via a script (macruby_install_xcode_support)
which is installed into /usr/local/bin along with the other scripts.

This allows the integration to be reinstalled easily. The script also has
reworked the install package so that it does not fail when Xcode integration
cannot be installed.